refactor(core): move log retention/reading from src-tauri to tyutool-core
P1a of the core consolidation plan (docs/specs/2026-08-26-core-consolidation-design.md).

Move the pure std/serde log-file retention and reading logic out of
src-tauri/src/logs.rs into crates/tyutool-core/src/diagnostics.rs, following
the P0 precedent set by prune_log_files/LogRetention:

- MAX_TRACE_FILES, prune_trace_files (.trace retention)
- pick_active_log, tail_bytes, read_log_tail_impl (active-log tail read)
- validate_log_filename, read_named_log_impl, resolve_log_open_path
  (path-traversal-safe named-log access)
- collect_log_files, LogFileInfo, list_log_files_impl (log listing)

src-tauri now calls tyutool_core functions directly from its Tauri commands and from
gather_and_write_logs_zip; no forwarding wrappers were left behind.

Left in src-tauri, per the crate-boundary rule (single-frontend pure-std code
stays with that frontend): the whole editor-detection block (LogFileOpener,
detect_vscode, detect_sublime_text, and friends, open_log_path_with_opener)
only ever serves the GUI's open-in-editor feature. Also untouched: the Tauri
command functions themselves, DialogPathRegistry (Tauri state), and
BatchAuthTraceWriter, build_report_info, redact_log_content, write_logs_zip,
gather_and_write_logs_zip, which are staged for P1b (zip export needs a
feature-gated zip dependency).

All 22 existing tests covering the moved functions (including the
validate_log_filename path-traversal cases) moved into
tyutool-core diagnostics tests unchanged; src-tauri keeps the tests for what
it still owns (editor catalog, redaction, zip writing, DialogPathRegistry).

No new dependency added to tyutool-core; confirmed tyutool-serve still has no
zip in its dependency tree.
